requirements.txt
langchain_ibm
ibm_watsonx_ai
python-dotenv
